requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
pycryptodome==3.17
pydantic==1.10.7
requests==2.30.0
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
urllib3==2.0.2
uvicorn==0.22.0